Add files via upload
The dataset hr_data.csv contains sample of candidates that we part of a
recruitment process of particular client of ScaleneWorks. ScaleneWorks
supports several information technology (IT) companies in India with
their talent acquisition. One of the challenge they face is about 30% of
the candidates who accept the jobs offers, do not join the company. This
lead to huge loss of revenue and time as the companies initiate the
recruitment process again to fill the workforce demand. ScaleneWorks
want to find out if a model can be build to predict the likelihood of a
candidate joining the company. If the likelihood is high, then the
company can go ahead and offer the jobs to the candidates